convert history_power_td/2 to int

+++ algo/operator.py
@@ -44,7 +44,7 @@ class Operator(util.OperatorBase):
         self.buffer_len = int(buffer_len)
         self.replay_buffer = deque(maxlen=self.buffer_len)
         self.power_history = deque(maxlen=history_power_td) # For history_power_td=60000 the power history of the ~7 days is stored.
-        self.daylight_power_history = deque(maxlen=history_power_td/2)
+        self.daylight_power_history = deque(maxlen=int(history_power_td/2))
         self.history_modus = history_modus
 
         self.p_1 = int(p_1)
